Support both c and cpp generically
Build the source and object lists separately. When linking just take
a wildcard list of objs in the build directory (I couldn't find a
nicer way to do it based on the proper object list without warnings
when we have foo.c and foo.cpp sources -> two copies of foo.obj in
the list and hence a warning). Still not perfect as we only end up
with one of the objects. I want to try and do something with prefixing
but I haven't made it work yet.